+++ test/bench/speed/tb_scr_fftw.c
@@ -64,7 +64,7 @@ main(int argc, char* argv[]) {
     if (argc > 1) {
 
         double ivals[TEST_COUNT];
-        size_t size = atoi(argv[1]);
+        size_t size = atol(argv[1]);
 
         fftw_scalar* fft_spectr =
             FFTW(malloc)(size * (sizeof(FFTW(complex)) + sizeof(fftw_scalar)));
